Add example controller test
Document the controller test case
