+++ src/controllers/example.js
@@ -1,0 +1,9 @@
+const db = require("../../database/db-config");
+
+module.exports = {
+  getAll
+};
+
+function getAll() {
+  return db.from("test").select("name");
+}
