--- src/controllers/nameTestModel.js
@@ -1,10 +1,0 @@
-const db = require('../../database/db-config');
-
-module.exports = {
-    getAll
-}
-
-function getAll() {
-    return db.from('test')
-    .select('name');
-}
